File Upload Function Created
The file upload function is created, but needs to be modified to accept auto incrementing and foreign keys as the ids.

Also to do the post requests I used axios, so please make sure to download that to your machines too.

Finally, I added csrf tokens to the pages for security.